DibApi.*:  remove dead files
