+++ GM/CBDsgn32.vcxproj
@@ -301,7 +301,6 @@ echo off
     <ClCompile Include="..\GShr\DragDrop.cpp" />
     <ClCompile Include="ClipBrd.cpp" />
     <ClCompile Include="..\GShr\Color.cpp" />
-    <ClCompile Include="..\GShr\DibApi.cpp" />
     <ClCompile Include="DlgBmask.cpp" />
     <ClCompile Include="DlgBrdp.cpp" />
     <ClCompile Include="DlgBrdsz.cpp" />
