+++ GM/CBDsgn32.vcxproj.filters
@@ -39,9 +39,6 @@
     <ClCompile Include="..\GShr\Color.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="..\GShr\DibApi.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
     <ClCompile Include="DlgBmask.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
