+++ GP/CBPlay32.vcxproj
@@ -298,7 +298,6 @@ echo off
     <ClCompile Include="..\GShr\CDib.cpp" />
     <ClCompile Include="..\GShr\CellForm.cpp" />
     <ClCompile Include="..\GShr\Color.cpp" />
-    <ClCompile Include="..\GShr\DibApi.cpp" />
     <ClCompile Include="..\GShr\DragDrop.cpp" />
     <ClCompile Include="DlgChgGameOwner.cpp" />
     <ClCompile Include="DlgDice.cpp" />
